Move platform event agnostic event utils into Miso.Event.
- [x] Hoist on, onWithOptions, and component mounting hooks into Miso.Event

This is a precursor refactor for native support.
